email-drive-validator/requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
Pillow>=10.0.0
python-dateutil>=2.8.0
